refactor(memory): refactor worker launch around source context & output deltas
