+++ EvoScientist/langgraph_dev/graphs.py
@@ -22,13 +22,11 @@ because it follows a different mechanism (re-exporting a lazily-constructed
 attribute), not the yaml-driven factory.
 """
 
-from EvoScientist.middleware.memory_lifecycle import (
-    MemoryLifecycleRole,
-    build_memory_worker_graph,
-)
+from EvoScientist.memory.agents import build_memory_worker_graph
+from EvoScientist.memory.types import MemorySourceType
 from EvoScientist.subagents._factory import build_async_subagent_graph
 
 writing_agent = build_async_subagent_graph("writing-agent")
 data_analysis_agent = build_async_subagent_graph("data-analysis-agent")
-evomemory_subagent_worker = build_memory_worker_graph(MemoryLifecycleRole.SUBAGENT)
-evomemory_turn_worker = build_memory_worker_graph(MemoryLifecycleRole.TURN)
+evomemory_subagent_worker = build_memory_worker_graph(MemorySourceType.SUBAGENT)
+evomemory_turn_worker = build_memory_worker_graph(MemorySourceType.TURN)
